Continue working to get generate_dem working

+++ geocal/lib/CartLab/vicar_file.cc
@@ -450,7 +450,7 @@ void VicarFile::label_set(const std::string& F,
 //-----------------------------------------------------------------------
 
 void VicarFile::label_set(const std::string& F, 
-			  double Val,
+			  float Val,
 			  const std::string& Property)
 {
 #ifdef HAVE_VICAR_RTL
